refactor: add index.ts to import files
